+++ packages/hono-takibi/src/generators/hono/openapi/component/generate-components-code.ts
@@ -3,7 +3,7 @@ import type { Config } from '../../../../config'
 import { generateZodToOpenAPISchemaDefinition } from '../../../zod/openapi/generate-zod-to-openapi-schema-definition'
 import { generateZodSchema } from '../../../zod/schema/generate-zod-schema'
 import { resolveSchemasDependencies } from '../../../../core/schema/references/resolve-schemas-dependencies'
-import { generateSchemasExport } from '../../../openapi/paths/generate-schemas-export'
+import { generateSchemasExport } from './schema/generate-schemas-export'
 import { getVariableSchemaNameHelper } from '../../../../core/helper/get-variable-schema-name-helper'
 
 /**
